Fixed linter error
